+++ hw2/include/point.h
@@ -4,7 +4,7 @@
 #include <istream>
 #include <glm/vec3.hpp>
 
-#define Point glm::dvec3
+#define Point glm::vec3
 
 class Ray {
 public:
@@ -14,10 +14,10 @@ public:
     Ray(const Point &o, const Point &d);
 
     Ray operator+(const Point &other) const;
-    Ray operator*(double k) const;
+    Ray operator*(float k) const;
 };
 
 std::istream& operator>>(std::istream &in, Point &p);
-Point operator*(double k, const Point &p);
+Point operator*(float k, const Point &p);
 
 #endif // DEFINE_POINT_H
