+++ hw2/src/color.cpp
@@ -3,40 +3,40 @@
 #include <cassert>
 #include <cmath>
 
-Color::Color::Color(double r, double g, double b): rgb(r, g, b) {}
-Color::Color::Color(glm::dvec3 rgb): rgb(rgb) {}
+Color::Color::Color(float r, float g, float b): rgb(r, g, b) {}
+Color::Color::Color(glm::vec3 rgb): rgb(rgb) {}
 
-double Color::r() const {
+float Color::r() const {
     return rgb.r;
 }
-double Color::g() const {
+float Color::g() const {
     return rgb.g;
 }
-double Color::b() const {
+float Color::b() const {
     return rgb.b;
 }
 
 static Color saturate(const Color &color) {
-    float r = std::max(std::min(1., color.r()), 0.);
-    float g = std::max(std::min(1., color.g()), 0.);
-    float b = std::max(std::min(1., color.b()), 0.);
+    float r = std::max(std::min(1.f, color.r()), 0.f);
+    float g = std::max(std::min(1.f, color.g()), 0.f);
+    float b = std::max(std::min(1.f, color.b()), 0.f);
     return {r, g, b};
 }
 
 Color AcesTonemap(const Color &col) {
     auto x = col.rgb;
 
-    const double a = 2.51f;
-    const double b = 0.03f;
-    const double c = 2.43f;
-    const double d = 0.59f;
-    const double e = 0.14f;
+    const float a = 2.51f;
+    const float b = 0.03f;
+    const float c = 2.43f;
+    const float d = 0.59f;
+    const float e = 0.14f;
     return saturate((x*(a*x+b))/(x*(c*x+d)+e));
 }
 
 Color GammaCorrected(const Color &col) {
     auto x = col.rgb;
-    double gamma = 1. / 2.2;
+    float gamma = 1. / 2.2;
     return {pow(x.r, gamma), pow(x.g, gamma), pow(x.b, gamma)};
 }
