+++ hw2/src/point.cpp
@@ -8,7 +8,7 @@ Ray::Ray(const Point &o, const Point &d): o(o), d(d) {}
 Ray Ray::operator+(const Point &p) const {
     return {o + p, d};
 }
-Ray Ray::operator*(double k) const {
+Ray Ray::operator*(float k) const {
     return {k * o, d};
 }
 
@@ -17,6 +17,6 @@ std::istream& operator>>(std::istream &in, Point &p) {
     return in;
 }
 
-Point operator*(double c, const Point &p) {
+Point operator*(float c, const Point &p) {
     return {c*p.x, c*p.y, c*p.z};
 }
